+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 
 - Updated mixpanel tracking
+- Input disabled when wallet is disconnected
 
 ## [1.0.7] - 2025-07-10